Change type for persistent ids.

+++ src/Todoist.Net/Models/IWithRelationsArgument.cs
@@ -5,6 +5,6 @@ namespace Todoist.Net.Models
 {
     internal interface IWithRelationsArgument
     {
-        void UpdateRelatedTempIds(IDictionary<Guid, int> map);
+        void UpdateRelatedTempIds(IDictionary<Guid, long> map);
     }
 }
